+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+- Initial release
+- Adds `StateFlow<T>` and `SharedFlow<T>`
+- Includes `FlowBuilder`, `PulseStreamBuilder`, and `FlowObserver`